+++ src/routes/character/new/enhancements/+page.server.ts
@@ -1,0 +1,9 @@
+import { enhancementManager } from "$lib/data/table-manager";
+const enhanceList = enhancementManager.getAll();
+import type { PageServerLoad } from './$types';
+
+export const load: PageServerLoad = async function() {
+  return {
+    list: enhanceList
+  };
+}